+++ lib/SKGB/Intern/Person.pm
@@ -3,20 +3,9 @@ package SKGB::Intern::Person;
 use 5.012;
 use utf8;
 use Carp qw( croak );
-use List::MoreUtils qw( none );
-use REST::Neo4p;
-use SKGB::Intern::Plugin::Neo4j;
 use Data::Dumper;
 
-
-my $Q = {
-  membership => REST::Neo4p::Query->new(<<QUERY),
-MATCH a=(p:Person)-[:ROLE|GUEST*]->(:Role {role:'member'})
- WHERE id(p) = {node}
- RETURN a
- LIMIT 1
-QUERY
-};
+use SKGB::Intern::Membership;
 
 
 sub new {
@@ -34,36 +23,6 @@ sub new {
 		return undef;
 	}
 	return bless $instance, $class;
-}
-
-
-sub new_membership {
-	my ($class, $neo4jPath) = @_;
-#	say Data::Dumper::Dumper $neo4jPath;
-	if ( ref $neo4jPath eq "REST::Neo4p::Path" ) {
-		$neo4jPath = $neo4jPath->as_simple;  # very slow due to REST::Neo4p::Path having to reload all entities from the server as their metadata is not in the original response (Neo4p uses the deprecated cypher endpoint of Neo4j)
-	}
-	elsif ( ref $neo4jPath eq "ARRAY" && scalar @$neo4jPath ) {
-		if ( ref $neo4jPath->[0] eq "REST::Neo4p::Node" ) {
-			# seems the 'fast' query was used
-			$neo4jPath = [
-				$neo4jPath->[0]->as_simple,
-				$neo4jPath->[1]->as_simple,
-				$neo4jPath->[2]->as_simple,
-			];
-		}
-		# else: do nothing (seems path query was used with ResponseAsObjects=0, which means the result is already in the format we need)
-	}
-	else {
-		croak "Neo4j path required";
-		return undef;
-	}
-	my $instance = bless {}, $class;
-#	my $instance = $class->new( ($neo4jPath->nodes)[0] );
-	$instance->{_simple} = $neo4jPath->[0];
-	$instance->membership($neo4jPath);
-#	say Data::Dumper::Dumper $instance;
-	return $instance;
 }
 
 
@@ -105,8 +64,6 @@ sub equals {
 sub node_id {
 	my ($self, $other) = @_;
 	return 0 + $self->{node}->{id} if $self->{node};
-#	return 0 + $self->{_simple}->{_node} if $self->{_simple};
-#	return 0 + $self->{_node}->id if $self->{_node};
 	die "node id unknown";
 }
 
@@ -221,150 +178,11 @@ sub gs_verein_id {
 
 sub membership {
 	my ($self, $path) = @_;
-	return $self->{membership} if $self->{membership};
-	
 	die "not implemented" if $path;
 	
-	$self->{membership} = { status_long => "Nichtmitglied" };
-	if (! defined $self->{relationships}) {
-		die "not implemented";
-		my $row = SKGB::Intern::Plugin::Neo4j::execute_memory($Q->{membership}, 1, (node => $self->node_id));
-		return $self->{membership} if ! $row->[0];
-		$path = $row->[0]->as_simple;
-	}
-	
-	my $date = POSIX::strftime('%Y-%m-%d', localtime);
-	my $joined = $self->_property('joined', relationship => undef);
-	my $leaves = $self->_property('leaves', relationship => undef);
-	return $self->{membership} if $joined && $joined gt $date || $leaves && $leaves lt $date;
-	
-	my $regular = $self->_property('regularContributor', relationship => undef);
-	my $status = $self->_property('role', node => undef);  # bug: matches any node
-	if ($status eq 'active-member') {
-		$status = "Aktiv"
-	}
-	elsif ($status eq 'passive-member') {
-		$status = "Passiv"
-	}
-	elsif ($status eq 'youth-member') {
-		$status = "Jugend"
-	}
-	elsif ($status eq 'honorary-member') {
-		$status = "Ehrenmitglied"
-	}
-	else {  # probably not a member
-		$status = $self->_property('name', node => undef);
-	}
-	$self->{membership}->{status} = $status;
-	$self->{membership}->{regular} = $regular;
-#	$self->{membership}->{guest} = $self->_property('_type', $path) eq 'IS_A_GUEST';
-	$self->{membership}->{guest} = grep {$_->{type} =~ m/^GUEST$/} @{$self->{relationships}};
-	$self->{membership}->{joined} = $joined;
-	$self->{membership}->{leaves} = $leaves;
-	
-	if ($self->_property('role', node => undef) eq 'honorary-member') {
-		$status = $regular ? "Aktiv, $status" : "Passiv, $status";
-	}
-	if ($self->{membership}->{guest}) {
-		$status .=  ", Gastmitglied";
-	}
-	$status and $self->{membership}->{status_long} = $status;
-#	say Dumper $self;
-#	die;
-	
+	$self->{membership} = SKGB::Intern::Membership->new( $self ) if ! $self->{membership};
 	return $self->{membership};
 }
-
-
-# sub membership_old {
-# 	my ($self) = @_;
-# 	return $self->{membership} if $self->{membership};
-# 	
-# 	$self->{membership} = { status_long => "Nichtmitglied" };
-# 	my $row = SKGB::Intern::Plugin::Neo4j::execute_memory($Q->{membership}, 1, (node => 0 + $self->{node}->id));
-# 	return $self->{membership} if ! $row->[0];
-# #	my @nodes = $row->[0]->nodes;
-# #	my @rels = $row->[0]->relationships;
-# 	my $path = $row->[0]->as_simple;
-# # 	while (my $n = shift @nodes) {
-# # 		my $r = shift @relns;
-# # 		print $r ? $n->id."-".$r->id."->" : $n->id."\n";
-# # 	}
-# 	
-# 	my $date = POSIX::strftime('%Y-%m-%d', localtime);
-# 	my $joined = $self->_property('joined', $path);
-# 	my $leaves = $self->_property('leaves', $path);
-# 	return $self->{membership} if $joined && $joined ge $date || $leaves && $leaves lt $date;
-# 	
-# 	my $regular = $self->_property('regularContributor', $path);
-# 	my $status = $self->_property('name', $path, 1);  # hack: path item 0 is the :Person
-# 	$self->{membership}->{status} = $status;
-# 	$self->{membership}->{regular} = $regular;
-# 	$self->{membership}->{guest} = $self->_property('_type', $path) eq 'IS_A_GUEST';
-# 	$self->{membership}->{joined} = $joined;
-# 	$self->{membership}->{leaves} = $leaves;
-# 	
-# 	if ($self->_property('role', $path) eq 'honorary-member') {
-# 		$status = $regular ? "Aktiv, $status" : "Passiv, $status";
-# 	}
-# 	if ($self->{membership}->{guest}) {
-# 		$status .=  ", Gastmitglied";
-# 	}
-# 	$self->{membership}->{status_long} = $status;
-# 	
-# 	return $self->{membership};
-# }
-# 
-# 
-# sub membership_oldest {
-# 	my ($self) = @_;
-# 	return $self->{membership} if $self->{membership};
-# 	
-# 	my @rels = $self->{node}->get_outgoing_relationships();
-# 	my $date = POSIX::strftime('%Y-%m-%d', localtime);
-# 	my ($status, $status_long, $membership);
-# 	foreach my $rel (@rels) {
-# 		next if $rel->type ne 'IS_A' && $rel->type ne 'IS_A_GUEST';
-# 		my $role = $rel->end_node->get_property('role');
-# 		next if $role eq 'user';  # TODO: hack; verify that this role really is a member role
-# #		say $self->name ." ". $role;
-# 		$membership = $rel;
-# 		
-# 		$status = $membership->end_node->get_property('name');
-# 		if ($role eq 'honorary-member') {
-# 			if ($membership->get_property('regularContributor')) {
-#  				$status_long = "Aktiv, $status";
-#  			}
-#  			else {
-#  				$status_long = "Passiv, $status";
-#  			}
-#  		}
-# 		if ($status && $membership->type eq 'IS_A_GUEST') {
-# 			$status .=  ", Gastmitglied";
-# 		}
-# 		$status_long ||= $status;
-# 		if ($membership->get_property('joined')) {
-# 			
-# 		}
-# 		my ($joined, $leaves) = ( $membership->get_property('joined'), $membership->get_property('leaves') );
-# 		$status = undef if $joined && $joined ge $date ||  $leaves && $leaves lt $date;
-# 		if (! $status) {
-# 			$status_long = "Nichtmitglied";
-# #			print $fh $mandate{umr} && $Gbetrag eq "0,00" ? "Nichtmitglied, Kontoinhaber" : "Nichtmitglied" if ! $status;
-# 		}
-# 		last;
-# 	}
-# 	
-# 	$self->{membership} = { status_long => $status_long };
-# 	if ($membership) {
-# 		$self->{membership}->{status} = $status;
-# 		$self->{membership}->{regular} = $membership->get_property('regularContributor');
-# 		$self->{membership}->{guest} = $membership->type eq 'IS_A_GUEST';
-# 		$self->{membership}->{joined} = $membership->get_property('joined');
-# 		$self->{membership}->{leaves} = $membership->get_property('leaves');
-# 	}
-# 	return $self->{membership};
-# }
 
  
 1;